requirements.txt
python-dotenv
aiogram==3.19.0
openai==1.70.0
deepseek==1.0.0